requirements.txt
pandas
streamlit
langchain
langchain-community
langchain-groq
langgraph
chromadb
sentence-transformers
pandas
python-dotenv